Modify board locations
Modified the gameboard locations to be objects
instead of arrays in order to gather more info
about each location and its status in the game
state. Added function that creates the initial
random ship location for the cpu ships. Added
location arrays to the ship factory and the board
factory in order to track locations that are
vacant or in use