Refactor funcs and add helper func
Added validlocationiterate helper func that more
cleanly and correctly checks to see if the
location inputs for the proposed ship are valid.
Refactored validattack func so that it checks to
see if the first part of the location input is
a string and then forces it to be uppercase.
Refactored placeship func so that it correctly
places the ship on the board when the letter value
of the location is iterating. Fixed validlocation
so that it properly iterates.