Modify ship inputs and location validators
Modified the ship constructor so that it allows
for directional input and for a name input.
Adjusted the placeship call to reflect this
change for more dynamic ship creation. Adjusted
validattack func so that it further checks to see
if the input will align with the allotted
letters in the alpha array. Fixed
validlocationhelperiterate so that it properly
uppercases the letter input.