Add randomizer and modify funcs
Added randomizer file and function that creates
random coordinates and a random direction for a
ship and then attempts to place that ship on the
board. Modified placeship so that it calls the
randomizer func if the ship generated is invalid.
Added hitlocations array in the ship constructor
in order to more accurately track which part of
the ships is hit.