Add length input options
Allowed the user to input ship length inputs and
added lists that contain the ship objects based
on their length. Added a function that checks to
see if the ship length inputted is valid based
on the allotted ship amounts in the ship lists.
Added functionality that pushes ships to their
respective lists.