Add ship and hit rendering
Created renderShips function that dynamically
renders the ships that are created to their
respective boards. Added renderHits function that
changes the color of a cell if a ship is present
there and is hit. Designed renderMissedAttack
function that changes a cell to yellow if it was
hit and there was no ship present there.