Add render and logic functions
Finished playerAttack function so that the attack
made by the human player would call the
receiveAttack function correctly. Created a
function for displaying when all the ships on a
board have been sunk and the game has ended.
Created a function that properly renders all the
click event functionality in order to allow the
human player to click where they want to attack.